build: accept any utils (#131)
# Motivation

The peer dependencies on `@dfinity/utils` in this linter is pain.
Everywhere we use it, either it throws error, block or throw warning
while fundamentally speaking we do not care if the linter inherits or
not that lib. So while it is not fully correct to require `*` since the
nullish utils were introduced in v0.0.1x something, we still set `*` as
peer to avoid being annoyed.