feat: Disable local rule `use-option-type-wrapper` by default (#138)
# Motivation

According to suggestion
https://github.com/dfinity/eslint-config-oisy-wallet/issues/135, it make
sense to disable local rule `use-option-type-wrapper` by default: it may
happen that the consumer does not want to create/use the `Option` type.

The consumer can always turn it on if necessary (updated documentation).

NOTE: local rule `use-nullish-checks` was under the same circumnstances,
but its metadata were not updated.