refactor: Split rules that are part of `recommended-type-checked` (#164)
# Motivation

I wanted to introduce the ruleset
[`recommended-type-checked`](https://typescript-eslint.io/users/configs/#recommended-type-checked)
for Typescript Eslint (see PR
https://github.com/dfinity/eslint-config-oisy-wallet/pull/162).

However, it is quite aggressive. Plus I would like to turn off at least
10 rules among them.

So, we take a different approach: I will include those rules on-demand,
one PR at the time. And when we have more than enough to justify that we
replace included rules with excluded ones, we change the ruleset.

In this PR, I just move the `recommended-type-checked` rules that we
already have in a different object, to manage them better.