chore: updates workflow to enable npm trusted publishing (#167)
# Motivation

NPM package publishing will be migrated to use Trusted Publishing (via
OIDC).

To ensure security and restrict access, we will configure the OIDC token
request to only be available to jobs running in the dedicated 'release'
environment. This environment will be a required condition for obtaining
the necessary NPM token.

# Changes

- Updates GitHub actions to latest versions
- Sets NPM version via engine configuration to >=11.5.1
- Merges both publish workflows into one workflow to comply with the
npmjs[.]org requirements